+++ NEWS.md
@@ -209,7 +209,7 @@
 # gmvarkit 2.1.0
 
 * Added support for recursively identified structural GSMVAR models.
-* Added the function linear_IRF to calculate linear impulse response functions based on a single regime. 
+* Added the function linear_IRF to calculate linear impulse response functions based on a single regime. Bootstrapped confidence bounds are also available for models that impose linear autoregressive dynamics. 
 * Added a possibility to constrain the mixing weight parameters alphas to fixed constants.
 * Added a possibility to constrain the lambda parameters of structural models to fixed constant.
 * fitGSMVAR can now filter out (many of the) inappropriate estimates by setting filter_estimates=TRUE.